Restructure tests and adjust autoconf
