+++ tests/integration/test.h
@@ -1,0 +1,2 @@
+
+extern bool g_runningInCI;
