+++ tests/manual/cmds.txt
