+++ tests/manual/compare_ops_in_two_folders.sh
