+++ tests/manual/purge_account.cpp
